consolidated more api endpoints
