--- src/app/utils/api/event.ts
@@ -1,5 +1,0 @@
-import { apiGet, apiPost } from "./api";
-import { TagType } from "../types";
-
-export const fetchTags = (eventId:number) =>
-  apiGet<TagType[]>(`/events/${eventId}/tags`);
